requirements.txt
flask
gunicorn
pymysql
python-dotenv
boto3
flask-cors
opencv-python
paho-mqtt
AWSIoTPythonSDK

